backend/requirements.txt
aiomysql==0.2.0
bcrypt==3.2.0
cryptography==40.0.2
fastapi==0.111.0
pycryptodome==3.20.0
passlib==1.7.4
python-jose==3.3.0
tortoise-orm==0.21.3
uvicorn==0.30.1
asyncpg==0.29.0
